:rotating_light: Fix Psalm issues With level 3

+++ src/Challenges/SURVIVAL_2/Human.php
@@ -89,7 +89,7 @@ final class Human {
 	 * @return void
 	 */
 	private function sleep() {
-		$this->energy += ( $this->thirst + $this->hunger ) / 2;
+		$this->energy += (int) ( ( $this->thirst + $this->hunger ) / 2 );
 		$this->thirst -= 5;
 		$this->hunger -= 5;
 	}
